requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.1.2
python-dotenv>=1.0.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
streamlit>=1.31.0
PyPDF2>=3.0.1
pdfplumber>=0.10.3
blinker>=1.7.0
huggingface-hub>=0.19.4
transformers>=4.36.2
datasets>=2.16.1
pandas>=2.2.0
torch>=2.2.1
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
PyMuPDF>=1.23.8
faiss-cpu>=1.7.4
nltk>=3.8.1
numpy>=1.24.3
scikit-learn>=1.3.2
tqdm>=4.66.1
regex>=2023.10.3
protobuf>=4.25.1
sentencepiece>=0.1.99
sacremoses>=0.0.53 